server/pyproject.toml
[tool.poetry]
name = "report-creator-ai"
version = "0.1.0"
description = ""
authors = ["Bruno Alves <nito.ba.dev@gmail.com>"]
readme = "README.md"

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
dev = 'python3 src/app.py'
api = 'fastapi dev src/http/app.py --host 0.0.0.0 --port 3333'
requires = 'poetry export --without-hashes --format=requirements.txt > requirements.txt'
migrate-up = 'alembic upgrade head'
pre_test = 'task lint'
test = 'pytest -s -x --cov=linkfy -vv'
post_test = 'coverage html'

[tool.poetry.dependencies]
python = "3.12.*"


pydantic-settings = "^2.3.1"
google-api-python-client = "^2.132.0"
google-auth = "^2.30.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
groq = "^0.9.0"
langchain-groq = "^0.1.5"
langchain-ollama = "^0.1.0rc0"
langchain = "^0.2.10"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.4.8"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 90
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
